+++ ChangeLog.md
@@ -2,7 +2,11 @@
 
 ## 0.1.1.0
 
-- Switch to types imported from `epi-types`.
+- Implement the likelihood for the birth-death-sampling model with inhomogeneous
+  birth rate making use of the functions in `BDSCOD.Llhd` to do the actual
+  computation.
+- Switch to types imported from `epi-types` and create a `BDSCOD.Types` module
+  for types shared between constant and variable rate likelihoods.
 
 ## 0.1.0.1